# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Calseta",
    "description": "SOC data platform for security agent consumption",
    "version": "dev"
  },
  "paths": {
...
    "/v1/alerts/{alert_uuid}/enrich": {
      "post": {
        "tags": [
          "alerts"
        ],
        "summary": "Enrich Alert",
        "description": "Re-trigger the enrichment pipeline for an alert.\n\nQueues the enrich_alert task which re-runs the full 3-pass indicator\nextraction + enrichment pipeline. Safe to call multiple times (idempotent).",
        "operationId": "enrich_alert_v1_alerts__alert_uuid__enrich_post",
        "parameters": [
          {
            "name": "alert_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Alert Uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/enrichments": {
      "post": {
        "tags": [
          "enrichments"
        ],
        "summary": "Enrich an indicator on-demand",
        "description": "Synchronously enrich an indicator against all configured providers.\n\nResults are returned cache-first. Each provider entry in the response\nincludes `cache_hit: true` when the result was served from cache rather\nthan a live API call.",
        "operationId": "enrich_on_demand_v1_enrichments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnDemandEnrichmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_OnDemandEnrichmentResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/enrichment-providers": {
      "get": {
        "tags": [
          "enrichment-providers"
        ],
        "summary": "List Enrichment Providers",
        "operationId": "list_enrichment_providers_v1_enrichment_providers_get",
        "parameters": [
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "name": "is_builtin",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Builtin"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (1-indexed)",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (1-indexed)"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Items per page (max 500)",
              "default": 50,
              "title": "Page Size"
            },
            "description": "Items per page (max 500)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_EnrichmentProviderResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "enrichment-providers"
        ],
        "summary": "Create Enrichment Provider",
        "operationId": "create_enrichment_provider_v1_enrichment_providers_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrichmentProviderCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_EnrichmentProviderResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/enrichment-providers/{provider_uuid}": {
      "get": {
        "tags": [
          "enrichment-providers"
        ],
        "summary": "Get Enrichment Provider",
        "operationId": "get_enrichment_provider_v1_enrichment_providers__provider_uuid__get",
        "parameters": [
          {
            "name": "provider_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Provider Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_EnrichmentProviderResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "enrichment-providers"
        ],
        "summary": "Patch Enrichment Provider",
        "operationId": "patch_enrichment_provider_v1_enrichment_providers__provider_uuid__patch",
        "parameters": [
          {
            "name": "provider_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Provider Uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrichmentProviderPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_EnrichmentProviderResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "enrichment-providers"
        ],
        "summary": "Delete Enrichment Provider",
        "operationId": "delete_enrichment_provider_v1_enrichment_providers__provider_uuid__delete",
        "parameters": [
          {
            "name": "provider_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Provider Uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/enrichment-providers/{provider_uuid}/test": {
      "post": {
        "tags": [
          "enrichment-providers"
        ],
        "summary": "Test Enrichment Provider",
        "operationId": "test_enrichment_provider_v1_enrichment_providers__provider_uuid__test_post",
        "parameters": [
          {
            "name": "provider_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Provider Uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrichmentProviderTestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_EnrichmentProviderTestResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/enrichment-providers/{provider_uuid}/activate": {
      "post": {
        "tags": [
          "enrichment-providers"
        ],
        "summary": "Activate Enrichment Provider",
        "operationId": "activate_enrichment_provider_v1_enrichment_providers__provider_uuid__activate_post",
        "parameters": [
          {
            "name": "provider_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Provider Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_EnrichmentProviderResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/enrichment-providers/{provider_uuid}/deactivate": {
      "post": {
        "tags": [
          "enrichment-providers"
        ],
        "summary": "Deactivate Enrichment Provider",
        "operationId": "deactivate_enrichment_provider_v1_enrichment_providers__provider_uuid__deactivate_post",
        "parameters": [
          {
            "name": "provider_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Provider Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_EnrichmentProviderResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataResponse_EnrichmentProviderResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EnrichmentProviderResponse"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DataResponse[EnrichmentProviderResponse]"
      },
      "DataResponse_EnrichmentProviderTestResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EnrichmentProviderTestResponse"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DataResponse[EnrichmentProviderTestResponse]"
      },
...
      "DataResponse_OnDemandEnrichmentResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/OnDemandEnrichmentResponse"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DataResponse[OnDemandEnrichmentResponse]"
      },
...
      "DataResponse_dict_": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DataResponse[dict]"
      },
...
      "EnrichmentProviderCreate": {
        "properties": {
          "provider_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-z0-9_]+$",
            "title": "Provider Name"
          },
          "display_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Display Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "supported_indicator_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Supported Indicator Types"
          },
          "http_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Http Config"
          },
          "auth_type": {
            "type": "string",
            "title": "Auth Type",
            "default": "no_auth"
          },
          "auth_config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Config"
          },
          "default_cache_ttl_seconds": {
            "type": "integer",
            "maximum": 86400.0,
            "minimum": 0.0,
            "title": "Default Cache Ttl Seconds",
            "default": 3600
          },
          "cache_ttl_by_type": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Ttl By Type"
          },
          "malice_rules": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Malice Rules"
          }
        },
        "type": "object",
        "required": [
          "provider_name",
          "display_name",
          "supported_indicator_types",
          "http_config"
        ],
        "title": "EnrichmentProviderCreate",
        "description": "Request body for POST /v1/enrichment-providers."
      },
...
      "EnrichmentProviderTestRequest": {
        "properties": {
          "indicator_type": {
            "type": "string",
            "title": "Indicator Type"
          },
          "indicator_value": {
            "type": "string",
            "title": "Indicator Value"
          }
        },
        "type": "object",
        "required": [
          "indicator_type",
          "indicator_value"
        ],
        "title": "EnrichmentProviderTestRequest",
        "description": "Request body for POST /v1/enrichment-providers/{uuid}/test."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IndicatorType": {
        "type": "string",
        "enum": [
          "ip",
          "domain",
          "hash_md5",
          "hash_sha1",
          "hash_sha256",
          "url",
          "email",
          "account"
        ],
        "title": "IndicatorType",
        "description": "Supported indicator-of-compromise (IOC) types."
      },
...
      "OnDemandEnrichmentRequest": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/IndicatorType"
          },
          "value": {
            "type": "string",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "type",
          "value"
        ],
        "title": "OnDemandEnrichmentRequest",
        "description": "Request body for POST /v1/enrichments.\n\nAccepts either {type, value} or {indicator_type, indicator_value}."
      },
...
      "PaginatedResponse_EnrichmentProviderResponse_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/EnrichmentProviderResponse"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "PaginatedResponse[EnrichmentProviderResponse]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}